Moved ProjectDependencyProvider to projects package
