+++ collect_app/src/main/java/org/odk/collect/android/backgroundwork/AutoSendTaskSpec.kt
@@ -19,10 +19,10 @@ import android.net.NetworkInfo
 import android.os.Environment
 import androidx.work.BackoffPolicy
 import androidx.work.WorkerParameters
-import org.odk.collect.android.formmanagement.ProjectDependencyProviderFactory
 import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.instancemanagement.InstanceAutoSender
 import org.odk.collect.android.network.NetworkStateProvider
+import org.odk.collect.android.projects.ProjectDependencyProviderFactory
 import org.odk.collect.android.utilities.FormsRepositoryProvider
 import org.odk.collect.async.TaskSpec
 import org.odk.collect.async.WorkerAdapter
